djangoServer/requirements.txt
attr==0.3.2
bpython==0.24
brotli==1.0.9
brotlicffi==1.0.9.2
ConfigParser==5.3.0
contextlib2==21.6.0
coreschema==0.0.4
cryptography==39.0.1
cx_Oracle==8.3.0
Cython==0.29.33
django_crispy_forms==2.0
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
ipdb==0.13.11
ipython==8.10.0
ipywidgets==8.0.4
jnius==1.1.0
keyring==23.13.1
lockfile==0.12.2
mock==5.0.1
MySQL-python==1.2.5
numpy==1.24.2
Pillow==9.4.0
protobuf==4.22.0
pylibmc==1.6.3
pymemcache==4.0.0
pyOpenSSL==23.0.0
pytidylib==0.3.2
pywatchman==1.4.1
PyYAML==6.0
railroad==0.5.0
sets==0.3.2
simplejson==3.18.3
Sphinx==6.1.3
tblib==1.7.0
tornado==6.2
trove_classifiers==2023.2.20
truststore==0.5.0
uritemplate==4.1.1
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
